requirement.txt
fastapi
uvicorn
requests
beautifulsoup4
readability-lxml
langdetect
